+++ packages/client/package.json
@@ -11,6 +11,7 @@
     "test": "jest ./"
   },
   "dependencies": {
+    "axios": "1.1.2",
     "dotenv": "16.0.2",
     "eslint-config-prettier": "8.5.0",
     "prettier": "2.7.1",
