+++ packages/client/src/App.tsx
@@ -1,8 +1,7 @@
 // import { useEffect } from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import PageNavigation from './components/page-navigation/page-navigation';
-import { ChangePassword } from './pages/change-password/change-password';
-import { Main, Login, SignUp, NewGame, Profile, Rules, Leaders, Forum, Round} from './pages/index';
+import { Main, Login, SignUp, NewGame, Profile, Rules, Leaders, Forum, Round, ChangePassword } from './pages/index';
 
 
 import './scss/style.scss';
@@ -29,11 +28,11 @@ export function App() {
         <Route path="/sign-up" element={<SignUp />} />
         <Route path="/profile" element={<Profile />} />
         <Route path="/change-password" element={<ChangePassword />} />
-        <Route path="/about" element={<Login/>}/>
-        <Route path="/new-game" element={<NewGame/>}/>
-        <Route path="/rules" element={<Rules/>}/>
-        <Route path="/leaders" element={<Leaders/>}/>
-        <Route path="/forum" element={<Forum/>}/>
+        <Route path="/about" element={<Login />} />
+        <Route path="/new-game" element={<NewGame />} />
+        <Route path="/rules" element={<Rules />} />
+        <Route path="/leaders" element={<Leaders />} />
+        <Route path="/forum" element={<Forum />} />
         <Route path="/round" element={<Round />} />
 
       </Routes>
