+++ packages/client/src/components/field/field.tsx
@@ -2,10 +2,11 @@ import React from 'react';
 import { UseFormRegisterReturn } from 'react-hook-form';
 import './field.scss';
 
-export interface IField {
+interface IField {
 	register: UseFormRegisterReturn<string>;
 	type?: string;
 	placeholder?: string;
+	value?:string
 }
 
 export function Field(props: IField) {
@@ -14,6 +15,7 @@ export function Field(props: IField) {
 			{...props.register}
 			type={props.type}
 			className={'field'}
+			value={props.value}
 			placeholder={props.placeholder}
 		/>
 	)
