+++ packages/client/src/components/form-field/form-field.scss
@@ -6,7 +6,7 @@
 
 .form-field--error {
   .field {
-    outline: 1px solid $alert;
+    border: 1px solid $alert;
   }
 
   &.form-field--error {
