+++ packages/client/src/components/intro/intro.scss
@@ -8,18 +8,12 @@
   justify-content: center;
   text-align: center;
   max-width: 420px;
-  margin: 0 auto;
+  margin: 0 auto 15px;
 }
 
 .intro--small {
   width: 100px;
   height: 100px;
-  margin: 0 auto 25px;
-
-  .intro__title,
-  .intro__desc {
-    display: none;
-  }
 }
 
 .intro__title {
@@ -35,7 +29,6 @@
 
 .intro__image-wrap {
   max-width: 300px;
-  margin-bottom: 30px;
 }
 
 @include down($mobile) {
