+++ packages/client/src/pages/index.tsx
@@ -7,3 +7,4 @@ export * from './rules';
 export * from './leaders';
 export * from './forum';
 export * from './round';
+export * from './change-password';
