+++ packages/client/src/scss/form/form.scss
@@ -39,7 +39,7 @@
 .form__buttons {
   display: flex;
   flex-direction: column;
-  gap: 28px;
+  gap: 24px;
 }
 
 .form__buttons {
